Set up webpack
The basic commands were set up to bundle the modules to make the
project work. The dist and src directories were created to store the
corresponding elements for the entry points and the output. On the other
hand, the style-loader and css-loader were locally installed to import
the css file to the html file in the dist folder.