Create DOM and navbar objects
These objects were created with factory functions and IIFE, as the app
only requires one object of each element. In the navbar object the
list property was defined, which is empty by default but is filled with
the DOM object method. Likewise, the navbar object contains the method
createList, which creates a new list element in the form of an object,
in which an ID with the current date occupies the first place, the name
of the element the second place, and an empty array that will contain
all the corresponding tasks of that section.

In the DOM object the method addNewListElement was defined, which
selects the respective form to create a new list element. These event
calls the addElemetList() function to create the element and append it
to the browser's respective position.