Save data in local storage
To do this the JSON.parsed was required when the property lists in the
navbar object was defined, so if there are elements in the local storage
they are called to fill the list, but if it is empty then the code
starts with an empty array.

On the other hand, in the DOM object the functions save and
saveAndRender were declared to store the current data. Finally, the
runOneTime method was defined in the DOM object to append the stored
elements in the respective place when the page is refreshed, otherwise
those elements will appear until the user adds a new one.