Make selected list different
A new property was defined in the navbar object, which is selectedID.
In order to select the desired list, an eventlistener was added to the
container of the lists, so when a LI element is clicked, its ID is taken
and the saveAndRender function is called to erase the whole list and
make it again, but this time the respective class with the style defined
is placed in the corresponding element. Likewise, the element selected
is stored in the local store, so when the page is refreshed, its style
remains.