Make some tasks appear in specific lists
The code leverages the fact that every list has an empty array for the
tasks, which should contain an id, the name of the task and if it is
completed or not. On the other hand, a template was created in the html
file, so the desired structure can be taken from there instead of having
to create all of that in JavaScript.

The respective code is an extension of the render funcion, which now
also call the new function renderTasks() thar oversees the creation
, to some extent at this point, of the tasks.